feat(nutritionlib): store macros as Float to capture decimal values
Fixes a bug where values like "Fat 0.4g" were silently dropped: the
integer-only regex matched "0", then the unit-guard rejected the row
because the next character was "." not "g".  Migrates calories, fat,
protein, and carbs from Int to Float throughout both modules (gramsPerServing
stays Int).  Also enables isReturnDefaultValues so readTextLines — which
calls android.util.Log — can be exercised in JVM unit tests.

+++ nutritionlib/src/main/java/com/graydyn/nutritionlib/model/Macros.kt
@@ -3,17 +3,17 @@ package com.graydyn.nutritionlib.model
 import java.io.Serializable
 
 data class Macros(
-    var calories: Int,
-    var fat: Int,
-    var protein: Int,
-    var carbs: Int,
+    var calories: Float,
+    var fat: Float,
+    var protein: Float,
+    var carbs: Float,
     var gramsPerServing: Int
 ) : Serializable {
 
-    constructor() : this(-1, -1, -1, -1, -1)
+    constructor() : this(-1f, -1f, -1f, -1f, -1)
 
     fun isComplete(proteinOnly: Boolean = false): Boolean {
-        if (proteinOnly) return calories != -1 && protein != -1
-        return calories != -1 && fat != -1 && protein != -1 && carbs != -1
+        if (proteinOnly) return calories != -1f && protein != -1f
+        return calories != -1f && fat != -1f && protein != -1f && carbs != -1f
     }
 }
